import random_u16 to sm64, match random_s32, and match actor func (#203)
* import random_u16 and atans from sm64

* rand math decomp

* Match func_80298AC0

+++ include/functions.h
@@ -3,7 +3,7 @@
 
 #include "common_structs.h"
 
-s32 random_int(s32);
+s32 random_int(u16);
 void mio0decode(u8* arg0, u8* arg1);
 
 void func_8000F2DC(void);
